Kubernetes

+++ install-debug.sh
@@ -563,14 +563,17 @@ then
 	# Docker-compose
 	sudo pip install docker-compose -y
 	
-	sudo apt-get update && sudo apt-get install -y apt-transport-https curl
-	sudo curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
-	cat <<EOF >/etc/apt/sources.list.d/kubernetes.list
-	deb https://apt.kubernetes.io/ kubernetes-xenial main
-	EOF
-	sudo apt-get update
-	sudo apt-get install -y kubelet kubeadm kubectl
-	sudo apt-mark hold kubelet kubeadm kubectl
+	read -p "Install Kubernetes ? yes or no" -n 1 -r
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+		sudo apt-get update && sudo apt-get install -y apt-transport-https curl
+		sudo curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
+		sudo cat <<EOF >/etc/apt/sources.list.d/kubernetes.list
+		deb https://apt.kubernetes.io/ kubernetes-xenial main
+		EOF
+		sudo apt-get update
+		sudo apt-get install -y kubelet kubeadm kubectl
+		sudo apt-mark hold kubelet kubeadm kubectl
 fi
 
 ########################################################################
